refactor: rename profitAndLossStorybook to profitAndLossStoryContext
