+++ src/components/ProfitAndLossDetailedCharts/ProfitAndLossDetailedCharts.stories.tsx
@@ -2,7 +2,7 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 
 import { ProfitAndLossDetailedCharts } from '@components/ProfitAndLossDetailedCharts/ProfitAndLossDetailedCharts'
 
-import { profitAndLossStoryHandlers, withProfitAndLossStoryContext } from '@test-utils/profitAndLossStorybook'
+import { profitAndLossStoryHandlers, withProfitAndLossStoryContext } from '@test-utils/profitAndLossStoryContext'
 
 type ProfitAndLossDetailedChartsStoryArgs = {
   scope: 'revenue' | 'expenses'
